lib : specify ods style enum
